+++ src/ExpressionLanguage/StringExpressionLanguageProvider.php
@@ -19,6 +19,11 @@ class StringExpressionLanguageProvider implements ExpressionFunctionProviderInte
 
                 return strlen($str);
             }),
+            new ExpressionFunction('contains', function ($haystack, $needle) {
+                return sprintf('(strpos(%1$s, %2$s) !== false) ? true : false', $haystack, $needle);
+            }, function ($arguments, $haystack, $needle) {
+                return strpos($haystack, $needle) !== false;
+            }),
         ];
     }
 }